refactor: Rename routes -> endpoints. Moved middleware folder.
