+++ backend/app/api/v1/endpoints/image.py
