+++ backend/app/api/v1/middleware/auth_middleware.py
