+++ backend/app/services/chat_service.py
@@ -5,7 +5,7 @@ from typing import Any, AsyncGenerator, Dict, Optional
 from app.config.loggers import chat_logger as logger
 from app.config.model_pricing import calculate_token_cost
 from app.langchain.core.agent import call_agent
-from app.middleware.tiered_rate_limiter import tiered_limiter
+from app.api.v1.middleware.tiered_rate_limiter import tiered_limiter
 from app.models.chat_models import (
     MessageModel,
     ToolDataEntry,
